VMM/DBGFCoreWrite: Windows build fix.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@32222 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/DBGFCoreWrite.cpp
@@ -169,7 +169,7 @@ static int Elf64WriteProgHdr(RTFILE hFile, uint32_t Type, uint32_t fFlags, uint6
  *
  * @return The size of the NOTE section as rounded to the file alignment.
  */
-static inline int Elf64NoteSectionSize(uint64_t cb)
+static inline uint64_t Elf64NoteSectionSize(uint64_t cb)
 {
     return sizeof(ELFNOTEHDR) + RT_ALIGN_64(cb, s_NoteAlign);
 }
